Centralize configuration
This attempts to centralize configuration definition and
discovery. This puts the configuration in one location and defines a
default source where the configuration comes from, environment
variables (lowest common denominator).

This also provides space for the definition of configuration from
other interfaces. For example, one could override these settings by on
the command-line interface (CLI) through CLI options. The implementer
of such a strategy is free to use whatever options they like as long
as under the hood the option's values conform to the settings defined
in this module.