Add a thread-local manager
This provides us with a framework agnostic bit of logic to place our
globals as thread-locals. Essentially, any application needs to have
one or more global bits, this providess a central location to place
those bits in a stacked format.

This is a paraphrased form of the threadlocal module found in the
Pyramid framework. I've gone this route, because the thread local
solution provided by werkzeug is specifically tailored for web server
use. The Pyramid solution is generalized enough--and I've generalized
it even more--to work in just about any situation.