Add a scripting preparation function
This loads the settings and application controller(s). The intended usage
is to be used as a contextual factory within a script, including the
main runtime script. The following is an example of what might show up
inside the body of a script's main function:

```python
with prepare(optional_settings_from_argv) as app_context:
    print(app_context['db']metadata.database.version)
```

Note, the application context's is bound to change. For now we're
database focused. Later we may have a better application startup
controller in place of some of the other minor controllers.