Add wbia-cnn to runtime requirements
`wbia-cnn` adds `generate_species_background_mask` to `IBEISController`
which seems to be central to a lot of the functionality in wildbook-ia.

We have 13 tests that don't work without `wbia-cnn`.  It seems that
`wbia-cnn` isn't like other plugins and should be installed as a runtime
requirement.

`wbia-cnn` requires `Lasagne` but the version on pypi is too old so need
the github version instead.  This has already been done in `wbia-cnn`
`requirements/runtime.txt` but it doesn't work because when the way
requirements are parsed.. or something.

Also setup.py `extras_require` doesn't accept "git+..." as valid
requirements:

```
error in wbia_cnn setup command: 'extras_require' must be a
dictionary whose values are strings or lists of strings containing
valid project/version requirement specifiers.
```

So adjust `parse_requirements` in `setup.py` to remove the git urls.