Deprecate the workdir concept in favor of env var settings
I'm moving us toward configuring the runtime in one central
location using the lowest common denominator, environment
variables. This implementation does provide a way for the developer to
still accept CLI arguments and options, but that is up to the
developer of that interface.

The workdir concept is closely tied with how the database is
initialized. These changes needed to make adjustments to the database
connection code. The preferred way of making a database connection is
now to use the URI (`DB_URL | db.runtime.url` in our settings).

Because the workdir functions are pretty widespread throughout the
code, I've only deprecated them and replaced their logic to point at
our new settings.