Condense scripting preparation into a forward compat func
The `wbia.scripting.prepare` prepare function is my hopeful attempt to
pull together a central location where all settings can be obtained at
startup. Right now it doesn't do much.

The intention is to begin steps towards determining the settings
before initializing the controller.

There is no promise that these changes are non-breaking. I've needed
to give up on non-breaking changes to get the central source of
configuration idea in place. For now it's in a halfway state, which is
better than prior attempts.