Integrate with DBAPIs on import of wbia
This is a bit of a hack, because we don't have a good way to load
code. If there were an entry-point that scanned for code integration,
then I'd use that, but it doens't exist (yet!).

This moves the integrations to the top-level package to guarantee they
get loaded.

I've stamped out a location for integrating psycopg2. Hopefully this
explains why I put this within a sub-package.